backend/requirements.txt
anthropic==0.94.1
flask==2.2.2
python-dotenv==1.2.2
